Introduce ErrorMessage component. Unify password prompt on BaseModal
