+++ lib/utils/base_mfa_dialog_state.dart
@@ -6,11 +6,11 @@ import 'package:flutter/material.dart';
 import 'constants.dart';
 
 abstract class BaseDialogState<T extends StatefulWidget> extends State<T> {
-  
+  final passwordField = TextEditingController();
   List<Widget> get dialogNext => [];
   Widget get dialogBody;
   int pageIndex = 0;
-  String errMsg = '';
+  String errorMessage = '';
   bool obscurePassword = true;
   bool inFlightRequest = false;
 
@@ -20,6 +20,20 @@ abstract class BaseDialogState<T extends StatefulWidget> extends State<T> {
   late String oobCode = '';
   late String mfaCode = '';
   bool requireAdditionalAuthentication = false;
+
+  @override
+  void initState() {
+    super.initState();
+    passwordField.addListener(() {
+      setState(() {});
+    });
+  }
+
+  @override
+  void dispose() {
+    passwordField.dispose();
+    super.dispose();
+  }
 
   Widget get dialogClose => IconButton(
     alignment: Alignment.centerLeft,
@@ -49,6 +63,59 @@ abstract class BaseDialogState<T extends StatefulWidget> extends State<T> {
     }
   }
 
+  Widget passwordPrompt(final String promptText, final VoidCallback onSubmit) => modalBody(
+    Align(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Text(
+            promptText,
+            textAlign: TextAlign.center,
+            softWrap: true,
+          ),
+          const SizedBox(height: 15),
+          SizedBox(
+            key: const Key('passwordField'),
+            width: 250,
+            child: TextFormField(
+              autofocus: true,
+              controller: passwordField,
+              onFieldSubmitted: (final value) {
+                onSubmit();
+              },
+              obscureText: obscurePassword,
+              enableSuggestions: false,
+              autocorrect: false,
+              autovalidateMode: AutovalidateMode.onUserInteraction,
+              validator: (final value) {
+                if (value == null || value.trim().isEmpty) {
+                  return 'Password is required';
+                }
+                return null;
+              },
+              decoration: InputDecoration(
+                suffixIcon: IconButton(
+                  key: const Key('toggle_password'),
+                  onPressed: () {
+                    setState(() {
+                      obscurePassword = !obscurePassword;
+                    });
+                  },
+                  icon: Icon(
+                      obscurePassword ? Icons.visibility : Icons.visibility_off
+                  ),
+                )
+              ),
+            ),
+          ),
+          const SizedBox(height: 15),
+          if (errorMessage.isNotEmpty)
+            Text(errorMessage, style: TextStyle(color: const ColorScheme.light().error))
+        ],
+      ),
+    )
+  );
+
   Widget modalBody(final Widget body) => Column(
     mainAxisSize: MainAxisSize.min,
     crossAxisAlignment: CrossAxisAlignment.start,
@@ -62,8 +129,6 @@ abstract class BaseDialogState<T extends StatefulWidget> extends State<T> {
         ),
     ],
   );
-
-  
   
   @override
   Widget build(final BuildContext context) {
